refactor(assets): colocate network filter label hook and reuse in control bar
Move useNetworkFilterButtonLabel to ui/components/app/assets/hooks and
use it from the home asset list control bar to remove duplicated label
logic.
