+++ ui/components/app/assets/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as usePrimaryCurrencyProperties } from './usePrimaryCurrencyProperties';
 export { default as useTokenDisplayInfo } from './useTokenDisplayInfo';
+export { useNetworkFilterButtonLabel } from './useNetworkFilterButtonLabel';
